Fix delete branch, category

+++ src/main/java/com/app/cosmetics/api/BranchApi.java
@@ -73,6 +73,9 @@ public class BranchApi {
 
     @DeleteMapping(path = "/{id}")
     public ResponseEntity<Void> deleteById(@PathVariable Long id) {
+        if (!authorizationService.isAdmin()) {
+            throw new NoAuthorizationException();
+        }
         branchService.delete(id);
         return ResponseEntity.ok().build();
     }
